+++ verik-importer/src/main/kotlin/io/verik/importer/cast/common/CasterVisitor.kt
@@ -76,11 +76,11 @@ class CasterVisitor(
     }
 
     override fun visitClassMethodFunction(ctx: SystemVerilogParser.ClassMethodFunctionContext?): SvElement? {
-        return FunctionCaster.castTaskFromClassMethodFunction(ctx!!, castContext)
+        return FunctionCaster.castFunctionFromClassMethodFunction(ctx!!, castContext)
     }
 
     override fun visitClassMethodExternMethod(ctx: SystemVerilogParser.ClassMethodExternMethodContext?): SvElement? {
-        return null
+        return FunctionCaster.castFunctionFromClassMethodExternMethod(ctx!!, castContext)
     }
 
     override fun visitClassMethodConstructor(ctx: SystemVerilogParser.ClassMethodConstructorContext?): SvElement? {
@@ -90,7 +90,7 @@ class CasterVisitor(
     override fun visitClassMethodExternConstructor(
         ctx: SystemVerilogParser.ClassMethodExternConstructorContext?
     ): SvElement? {
-        return null
+        return ConstructorCaster.castConstructorFromClassMethodExternConstructorContext(ctx!!, castContext)
     }
 
     override fun visitClassConstructorDeclaration(
@@ -184,7 +184,7 @@ class CasterVisitor(
     }
 
     override fun visitFunctionPrototype(ctx: SystemVerilogParser.FunctionPrototypeContext?): SvElement? {
-        return null
+        return FunctionCaster.castFunctionFromFunctionPrototype(ctx!!, castContext)
     }
 
 // A.2.7 Task Declarations /////////////////////////////////////////////////////////////////////////////////////////////
@@ -195,13 +195,13 @@ class CasterVisitor(
 
     override fun visitTaskBodyDeclarationNoPortList(
         ctx: SystemVerilogParser.TaskBodyDeclarationNoPortListContext?
-    ): SvElement {
+    ): SvElement? {
         return TaskCaster.castTaskFromTaskBodyDeclarationNoPortList(ctx!!, castContext)
     }
 
     override fun visitTaskBodyDeclarationPortList(
         ctx: SystemVerilogParser.TaskBodyDeclarationPortListContext?
-    ): SvElement {
+    ): SvElement? {
         return TaskCaster.castTaskFromTaskBodyDeclarationPortList(ctx!!, castContext)
     }
 
@@ -209,8 +209,12 @@ class CasterVisitor(
         return ValueParameterCaster.castValueParameterFromTfPortItem(ctx!!, castContext)
     }
 
-    override fun visitTaskPrototype(ctx: SystemVerilogParser.TaskPrototypeContext?): SvElement {
-        return super.visitTaskPrototype(ctx)
+    override fun visitTfPortDeclaration(ctx: SystemVerilogParser.TfPortDeclarationContext?): SvElement? {
+        return ValueParameterCaster.castValueParameterFromTfPortDeclaration(ctx!!, castContext)
+    }
+
+    override fun visitTaskPrototype(ctx: SystemVerilogParser.TaskPrototypeContext?): SvElement? {
+        return TaskCaster.castTaskFromTaskPrototype(ctx!!, castContext)
     }
 
 // A.6.2 Procedural Blocks and Assignments /////////////////////////////////////////////////////////////////////////////