+++ verik-importer/src/test/kotlin/io/verik/importer/cast/cast/ConstructorCasterTest.kt
@@ -24,7 +24,20 @@ import org.junit.jupiter.api.Test
 internal class ConstructorCasterTest : BaseTest() {
 
     @Test
-    fun `cast constructor from classConstructorDeclarations`() {
+    fun `cast constructor from classMethodExternConstructor`() {
+        driveCasterTest(
+            SystemVerilogParser.ClassMethodExternConstructorContext::class,
+            """
+                class c;
+                    extern function new(logic x);
+                endclass
+            """.trimIndent(),
+            "Constructor([ValueParameter(x, Nothing, SimpleDescriptor(Boolean))])"
+        ) { it.findDeclaration("new") }
+    }
+
+    @Test
+    fun `cast constructor from classConstructorDeclaration`() {
         driveCasterTest(
             SystemVerilogParser.ClassConstructorDeclarationContext::class,
             """